requirements.txt
kubernetes
google-generativeai
requests
boto3
azure-identity
